Speed/Power tweaks
Hi,

I made it so it's more survival-like, and so in order to equalise the output of a normal Assembler/Refinery you would need more Tiny Assemblers/Refineries than what you would need to fill up the same space. And it would use more power.

Also, I reduced the build time in half... just because that time is usually for a full Refinery/Assembler.

 After all, you're supposed to use those in a minimal-survival type of play, right? If you update the mod to be DX11-compatible and you leave a message for the creators of [this ship](https://steamcommunity.com/sharedfiles/filedetails/?id=566017523&searchtext=hardcore+landing+) and other tiny ships, you might have enough popularity with it that it would become a mainstream mod for hard/hardcore survival servers/plays.

Good luck with it!